Move TensorArray class from mps.py to array.py.
Also modify toMPS in MPSSum to have more options for the truncation.
