+++ seemps/state/array.py
@@ -1,0 +1,52 @@
+class TensorArray(object):
+    """TensorArray class.
+
+    This class provides the basis for all tensor networks. The class abstracts
+    a one-dimensional array of tensors that is freshly copied whenever the
+    object is cloned. Two TensorArray's can share the same tensors and be
+    destructively modified.
+
+    Attributes:
+    size = number of tensors in the array
+    """
+
+    def __init__(self, data):
+        """Create a new TensorArray from a list of tensors. `data` is an
+        iterable object, such as a list or other sequence. The list is cloned
+        before storing it into this object, so as to avoid side effects when
+        destructively modifying the array."""
+        self._data = list(data)
+        self.size = len(self._data)
+
+    def __getitem__(self, k):
+        #
+        # Get MP matrix at position `k`. If 'A' is an MP, we can now
+        # do A[k]
+        #
+        return self._data[k]
+
+    def __setitem__(self, k, value):
+        #
+        # Replace matrix at position `k` with new tensor `value`. If 'A'
+        # is an MP, we can now do A[k] = value
+        #
+        self._data[k] = value
+        return value
+
+    def __copy__(self):
+        #
+        # Return a copy of the MPS with a fresh new array.
+        #
+        return type(self)(self._data)
+
+    def __iter__(self):
+        return self._data.__iter__()
+
+    def __len__(self):
+        return self.size
+
+    def copy(self):
+        """Return a fresh new TensorArray that shares the same tensor as its
+        sibling, but which can be destructively modified without affecting it.
+        """
+        return self.__copy__()
