+++ seemps/state/mps.py
@@ -1,63 +1,10 @@
 import copy
 import numpy as np
+from .array import TensorArray
 from .. import expectation
 from .schmidt import vector2mps
 from .truncation import DEFAULT_TOLERANCE
 import warnings
-
-
-class TensorArray(object):
-    """TensorArray class.
-
-    This class provides the basis for all tensor networks. The class abstracts
-    a one-dimensional array of tensors that is freshly copied whenever the
-    object is cloned. Two TensorArray's can share the same tensors and be
-    destructively modified.
-
-    Attributes:
-    size = number of tensors in the array
-    """
-
-    def __init__(self, data):
-        """Create a new TensorArray from a list of tensors. `data` is an
-        iterable object, such as a list or other sequence. The list is cloned
-        before storing it into this object, so as to avoid side effects when
-        destructively modifying the array."""
-        self._data = list(data)
-        self.size = len(self._data)
-
-    def __getitem__(self, k):
-        #
-        # Get MP matrix at position `k`. If 'A' is an MP, we can now
-        # do A[k]
-        #
-        return self._data[k]
-
-    def __setitem__(self, k, value):
-        #
-        # Replace matrix at position `k` with new tensor `value`. If 'A'
-        # is an MP, we can now do A[k] = value
-        #
-        self._data[k] = value
-        return value
-
-    def __copy__(self):
-        #
-        # Return a copy of the MPS with a fresh new array.
-        #
-        return type(self)(self._data)
-
-    def __iter__(self):
-        return self._data.__iter__()
-
-    def __len__(self):
-        return self.size
-
-    def copy(self):
-        """Return a fresh new TensorArray that shares the same tensor as its
-        sibling, but which can be destructively modified without affecting it.
-        """
-        return self.__copy__()
 
 
 class MPS(TensorArray):
@@ -495,17 +442,25 @@ class MPSSum:
         the complete wavefunction that is encoded in the MPS."""
         return sum(wa * A.to_vector() for wa, A in zip(self.weights, self.states))
 
-    def toMPS(self, normalize=None):
+    def toMPS(
+        self, normalize=None, tolerance=None, maxsweeps=16, max_bond_dimension=None
+    ):
         from ..truncate.combine import combine
 
         if normalize is None:
             normalize = self.normalize
+        if tolerance is None:
+            tolerance = self.tolerance
+        if maxsweeps is None:
+            maxsweeps = self.maxsweeps
+        if max_bond_dimension is None:
+            max_bond_dimension = self.max_bond_dimension
         ψ, _ = combine(
             self.weights,
             self.states,
-            maxsweeps=self.maxsweeps,
-            tolerance=self.tolerance,
+            maxsweeps=maxsweeps,
+            tolerance=tolerance,
             normalize=normalize,
-            max_bond_dimension=self.max_bond_dimension,
+            max_bond_dimension=max_bond_dimension,
         )
         return ψ
